// schema.graphql/API.ts
export type Business = {
  __typename: "Business",
  address: string,
  billingType?: BusinessBillingType | null,
  businessType?: BusinessBusinessType | null,
  city: string,
  cpmRate?: number | null,
  createdAt: string,
  currentSpend?: number | null,
  description?: string | null,
  email: string,
  id: string,
  isActive?: boolean | null,
  isVerified?: boolean | null,
  latitude?: number | null,
  logoKey?: string | null,
  logoUrl?: string | null,
  longitude?: number | null,
  monthlyBudget?: number | null,
  name: string,
  owner: string,
  phone?: string | null,
  rejectionReason?: string | null,
  rewards?: ModelRewardConnection | null,
  serviceRadius?: number | null,
  state: string,
  updatedAt: string,
  verificationStatus?: BusinessVerificationStatus | null,
  website?: string | null,
  zipCode: string,
};

export enum BusinessBillingType {
  CPC = "CPC",
  CPM = "CPM",
  MONTHLY_RETAINER = "MONTHLY_RETAINER",
}


export enum BusinessBusinessType {
  EDUCATION = "EDUCATION",
  ENTERTAINMENT = "ENTERTAINMENT",
  OTHER = "OTHER",
  RESTAURANT = "RESTAURANT",
  RETAIL = "RETAIL",
  SERVICES = "SERVICES",
  SPORTS_RECREATION = "SPORTS_RECREATION",
}


export type ModelRewardConnection = {
  __typename: "ModelRewardConnection",
  items:  Array<Reward | null >,
  nextToken?: string | null,
};
// ...
export enum RewardRedemptionType {
  CODE = "CODE",
  IN_STORE = "IN_STORE",
  ONLINE = "ONLINE",
  QR_CODE = "QR_CODE",
}
// ...
export enum BusinessVerificationStatus {
  APPROVED = "APPROVED",
  PENDING = "PENDING",
  REJECTED = "REJECTED",
}


export type BusinessAnalytics = {
  __typename: "BusinessAnalytics",
  avgChildAge?: number | null,
  businessId: string,
  clicks?: number | null,
  createdAt: string,
  date: string,
  id: string,
  impressions?: number | null,
  owner?: string | null,
  redemptions?: number | null,
  revenue?: number | null,
  spend?: number | null,
  topRewardId?: string | null,
  updatedAt: string,
};
// ...
export type GetBusinessQueryVariables = {
  id: string,
};

export type GetBusinessQuery = {
  getBusiness?:  {
    __typename: "Business",
    address: string,
    billingType?: BusinessBillingType | null,
    businessType?: BusinessBusinessType | null,
    city: string,
    cpmRate?: number | null,
    createdAt: string,
    currentSpend?: number | null,
    description?: string | null,
    email: string,
    id: string,
    isActive?: boolean | null,
    isVerified?: boolean | null,
    latitude?: number | null,
    logoKey?: string | null,
    logoUrl?: string | null,
    longitude?: number | null,
    monthlyBudget?: number | null,
    name: string,
    owner: string,
    phone?: string | null,
    rejectionReason?: string | null,
    rewards?:  {
      __typename: "ModelRewardConnection",
      nextToken?: string | null,
    } | null,
    serviceRadius?: number | null,
    state: string,
    updatedAt: string,
    verificationStatus?: BusinessVerificationStatus | null,
    website?: string | null,
    zipCode: string,
  } | null,
};

export type GetBusinessAnalyticsQueryVariables = {
  id: string,
};

export type GetBusinessAnalyticsQuery = {
  getBusinessAnalytics?:  {
    __typename: "BusinessAnalytics",
    avgChildAge?: number | null,
    businessId: string,
    clicks?: number | null,
    createdAt: string,
    date: string,
    id: string,
    impressions?: number | null,
    owner?: string | null,
    redemptions?: number | null,
    revenue?: number | null,
    spend?: number | null,
    topRewardId?: string | null,
    updatedAt: string,
  } | null,
};

export type GetRewardQueryVariables = {
  id: string,
};

export type GetRewardQuery = {
  getReward?:  {
    __typename: "Reward",
    availableRadius?: number | null,
    business?:  {
      __typename: "Business",
      address: string,
      billingType?: BusinessBillingType | null,
      businessType?: BusinessBusinessType | null,
      city: string,
      cpmRate?: number | null,
      createdAt: string,
      currentSpend?: number | null,
      description?: string | null,
      email: string,
      id: string,
      isActive?: boolean | null,
      isVerified?: boolean | null,
      latitude?: number | null,
      logoKey?: string | null,
      logoUrl?: string | null,
      longitude?: number | null,
      monthlyBudget?: number | null,
      name: string,
      owner: string,
      phone?: string | null,
      rejectionReason?: string | null,
      serviceRadius?: number | null,
      state: string,
      updatedAt: string,
      verificationStatus?: BusinessVerificationStatus | null,
      website?: string | null,
      zipCode: string,
    } | null,
    businessId: string,
    category?: RewardCategory | null,
    clicks?: number | null,
    createdAt: string,
    currentRedemptions?: number | null,
    dailyLimit?: number | null,
    daysOfWeek?: Array< string | null > | null,
    description: string,
    endDate?: string | null,
    id: string,
    imageKey?: string | null,
    imageUrl?: string | null,
    impressions?: number | null,
    isActive?: boolean | null,
    owner?: string | null,
    pointCost: number,
    redemptionInstructions?: string | null,
    redemptionRecords?:  {
      __typename: "ModelRewardRedemptionConnection",
      nextToken?: string | null,
    } | null,
    redemptionType?: RewardRedemptionType | null,
    redemptions?: number | null,
    retailValue?: number | null,
    rewardType?: RewardRewardType | null,
    startDate: string,
    status?: RewardStatus | null,
    targetAgeRanges?: Array< string | null > | null,
    targetFamilySizes?: Array< number | null > | null,
    termsAndConditions?: string | null,
    title: string,
    totalAvailable?: number | null,
    updatedAt: string,
  } | null,
};

export type GetRewardRedemptionQueryVariables = {
  id: string,
};

export type GetRewardRedemptionQuery = {
  getRewardRedemption?:  {
    __typename: "RewardRedemption",
    businessId?: string | null,
    businessRevenue?: number | null,
    childAge?: number | null,
    childId: string,
    createdAt: string,
    expiresAt?: string | null,
    familyId: string,
    fulfilledAt?: string | null,
    id: string,
    notes?: string | null,
    owner?: string | null,
    pointsSpent: number,
    redemptionCity?: string | null,
    redemptionCode?: string | null,
    redemptionDate?: string | null,
    redemptionStatus?: RewardRedemptionRedemptionStatus | null,
    redemptionZip?: string | null,
    reward?:  {
      __typename: "Reward",
      availableRadius?: number | null,
      businessId: string,
      category?: RewardCategory | null,
      clicks?: number | null,
      createdAt: string,
      currentRedemptions?: number | null,
      dailyLimit?: number | null,
      daysOfWeek?: Array< string | null > | null,
      description: string,
      endDate?: string | null,
      id: string,
      imageKey?: string | null,
      imageUrl?: string | null,
      impressions?: number | null,
      isActive?: boolean | null,
      owner?: string | null,
      pointCost: number,
      redemptionInstructions?: string | null,
      redemptionType?: RewardRedemptionType | null,
      redemptions?: number | null,
      retailValue?: number | null,
      rewardType?: RewardRewardType | null,
      startDate: string,
      status?: RewardStatus | null,
      targetAgeRanges?: Array< string | null > | null,
      targetFamilySizes?: Array< number | null > | null,
      termsAndConditions?: string | null,
      title: string,
      totalAvailable?: number | null,
      updatedAt: string,
    } | null,
    rewardId: string,
    updatedAt: string,
  } | null,
};

export type ListBusinessAnalyticsQueryVariables = {
  filter?: ModelBusinessAnalyticsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListBusinessAnalyticsQuery = {
  listBusinessAnalytics?:  {
    __typename: "ModelBusinessAnalyticsConnection",
    items:  Array< {
      __typename: "BusinessAnalytics",
      avgChildAge?: number | null,
      businessId: string,
      clicks?: number | null,
      createdAt: string,
      date: string,
      id: string,
      impressions?: number | null,
      owner?: string | null,
      redemptions?: number | null,
      revenue?: number | null,
      spend?: number | null,
      topRewardId?: string | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListBusinessesQueryVariables = {
  filter?: ModelBusinessFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListBusinessesQuery = {
  listBusinesses?:  {
    __typename: "ModelBusinessConnection",
    items:  Array< {
      __typename: "Business",
      address: string,
      billingType?: BusinessBillingType | null,
      businessType?: BusinessBusinessType | null,
      city: string,
      cpmRate?: number | null,
      createdAt: string,
      currentSpend?: number | null,
      description?: string | null,
      email: string,
      id: string,
      isActive?: boolean | null,
      isVerified?: boolean | null,
      latitude?: number | null,
      logoKey?: string | null,
      logoUrl?: string | null,
      longitude?: number | null,
      monthlyBudget?: number | null,
      name: string,
      owner: string,
      phone?: string | null,
      rejectionReason?: string | null,
      serviceRadius?: number | null,
      state: string,
      updatedAt: string,
      verificationStatus?: BusinessVerificationStatus | null,
      website?: string | null,
      zipCode: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListRewardRedemptionsQueryVariables = {
  filter?: ModelRewardRedemptionFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListRewardRedemptionsQuery = {
  listRewardRedemptions?:  {
    __typename: "ModelRewardRedemptionConnection",
    items:  Array< {
      __typename: "RewardRedemption",
      businessId?: string | null,
      businessRevenue?: number | null,
      childAge?: number | null,
      childId: string,
      createdAt: string,
      expiresAt?: string | null,
      familyId: string,
      fulfilledAt?: string | null,
      id: string,
      notes?: string | null,
      owner?: string | null,
      pointsSpent: number,
      redemptionCity?: string | null,
      redemptionCode?: string | null,
      redemptionDate?: string | null,
      redemptionStatus?: RewardRedemptionRedemptionStatus | null,
      redemptionZip?: string | null,
      rewardId: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};

export type ListRewardsQueryVariables = {
  filter?: ModelRewardFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListRewardsQuery = {
  listRewards?:  {
    __typename: "ModelRewardConnection",
    items:  Array< {
      __typename: "Reward",
      availableRadius?: number | null,
      businessId: string,
      category?: RewardCategory | null,
      clicks?: number | null,
      createdAt: string,
      currentRedemptions?: number | null,
      dailyLimit?: number | null,
      daysOfWeek?: Array< string | null > | null,
      description: string,
      endDate?: string | null,
      id: string,
      imageKey?: string | null,
      imageUrl?: string | null,
      impressions?: number | null,
      isActive?: boolean | null,
      owner?: string | null,
      pointCost: number,
      redemptionInstructions?: string | null,
      redemptionType?: RewardRedemptionType | null,
      redemptions?: number | null,
      retailValue?: number | null,
      rewardType?: RewardRewardType | null,
      startDate: string,
      status?: RewardStatus | null,
      targetAgeRanges?: Array< string | null > | null,
      targetFamilySizes?: Array< number | null > | null,
      termsAndConditions?: string | null,
      title: string,
      totalAvailable?: number | null,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};